Rearange the tuning stuff
